[refactor] Move semester constants to separate constants directory and update imports
