+++ src/constants/semester-constants.ts
