+++ src/sections/mileage-scholarship-application/course-completion-form.tsx
@@ -15,10 +15,9 @@ import {
   FormControlLabel,
 } from '@mui/material';
 
-import { SEMESTER_OPTIONS } from 'src/utils/semester-constants';
-
 import { useTranslate } from 'src/locales';
 import { Course } from 'src/domain/mileage-management/course';
+import { SEMESTER_OPTIONS } from 'src/constants/semester-constants';
 
 const GRADE_OPTIONS = ['A+', 'A0', 'A-', 'B+', 'PD', 'P'];
